messages: notifications,outbox,marking as read,deleting

+++ src/main/java/gr/uoa/di/tedi/projectbackend/handling/MessageNotFoundException.java
@@ -1,5 +1,5 @@
 package gr.uoa.di.tedi.projectbackend.handling;
 
 public class MessageNotFoundException extends RuntimeException{
-    public MessageNotFoundException(Long id){super("Could not find Seller" + id);}
+    public MessageNotFoundException(Integer id){super("Could not find Seller" + id);}
 }
