+++ src/main/java/gr/uoa/di/tedi/projectbackend/repos/MessageRepository.java
@@ -4,5 +4,5 @@ import gr.uoa.di.tedi.projectbackend.model.Location;
 import gr.uoa.di.tedi.projectbackend.model.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<Message,Long> ,CustomMessageRepository{
+public interface MessageRepository extends JpaRepository<Message,Integer> ,CustomMessageRepository{
 }
